Move key events out of index.js
